applespi: use debugfs for debugging touchpad dimensions.
Instead of enabling this via another flag in the 'debug' param and
showing the results in kernel log, we now use debugfs for this. Under
$debugfs-mnt/applespi (typically /sys/kernel/debug/applespi/) we create
two debugfs entries, 'enable_tp_dim' which enables/disables the tracing
of touchpad dimensions, and 'tp-dim' that shows the current values.

From the review by Andy Shevchenko and Greg Kroah-Hartman.