applespi: Fix rare NULL pointer derefence.
When an error is returned from the write we will try to schedule the
next write command. If at this time a read command is already scheduled,
we could end up re-using the current completion slot. But because we
were clearing the slot afterwards again, this would then lead to a NULL
pointer dereference when the next command completed.

Fix this by clearing the slot before calling the completion handler.

Signed-off-by: Ronald Tschalär <ronald@innovation.ch>